Refactor terramonitor
