+++ src/components/Map/layers/common/Terramonitor/satellite.ts
@@ -1,0 +1,35 @@
+import {
+  LayerGroupId,
+  LayerConf,
+  ExtendedStyleSpecification,
+} from '#/common/types/map'
+
+const id: LayerGroupId = 'terramonitor'
+
+const getStyle = async (): Promise<ExtendedStyleSpecification> => {
+  return {
+    version: 8,
+    name: id,
+    sources: {
+      [id]: {
+        type: 'raster',
+        tiles: [
+          `https://tm2.terramonitor.com/${process.env.NEXT_PUBLIC_TERRAMONITOR_KEY}/rgb/{z}/{x}/{y}.png`,
+        ],
+        tileSize: 256,
+        attribution:
+          '<a href="https://www.terramonitor.com">© Terramonitor</a>',
+      },
+    },
+    layers: [
+      {
+        id: 'terramonitor',
+        source: 'terramonitor',
+        type: 'raster',
+        paint: {},
+      },
+    ],
+  }
+}
+
+export const layerConf: LayerConf = { id: id, style: getStyle }
