--- src/components/Map/layers/extras/Terramonitor/index.ts
@@ -1,3 +1,0 @@
-import layerConf from './layerConf'
-
-export default layerConf
